+++ src/constants/base.js
@@ -5,4 +5,5 @@ export const columnWidths = {
   [searchResultListColumns.AUTH_REF_TYPE]: { min: 140, max: 150 },
   [searchResultListColumns.HEADING_REF]: { min: 370, max: 370 },
   [searchResultListColumns.HEADING_TYPE]: { min: 120, max: 140 },
+  [searchResultListColumns.AUTHORITY_SOURCE]: '250px',
 };